[feature/3] 로그인 시 리다이렉트, 콘텐츠 그래프 작업
