+++ src/views/Admin/styles.css
@@ -57,7 +57,6 @@
 .delete-button,
 .save-button {
   font-family: 'UhBeeSe_hyun';
-
   background-color: #F0E4F6;
   color: #555;
   border: none;
